backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
chromadb==0.5.0
sentence-transformers==2.7.0
requests==2.31.0
pydantic==2.7.0
